+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/ParadoxLocalisationCompletionContributor.kt
@@ -10,6 +10,7 @@ import icu.windea.pls.*
 import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.localisation.psi.ParadoxLocalisationTypes.*
 import icu.windea.pls.script.codeInsight.*
+import icu.windea.pls.script.psi.*
 
 class ParadoxLocalisationCompletionContributor : CompletionContributor() {
 	companion object{